pyproject.toml
[tool.poetry]
name = "pykit"
version = "0.1.0"
description = ""
authors = ["wwqdrh <huiloademail@163.com>"]
readme = "README.md"
packages = [{include = "pykit/logger"}]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
pyyaml = "^6.0"
pydantic = "^1.10.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
mypy = "^0.991"


[tool.poetry.group.types.dependencies]
types-pyyaml = "^6.0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
